fix: resolve PR lint failures for scoped factory document save
Split oversized scoped-save and graph-editor modules, route validation
target summary copy through the graph editor message catalog, and align
shared test helpers with dashboard session import boundaries.
