+++ ui/src/features/current-selection/base/hooks/useScopedFactoryDocumentSave.test.tsx
@@ -1,6 +1,4 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { act, renderHook, waitFor } from "@testing-library/react";
-import type { ReactNode } from "react";
 
 import { CurrentFactoryDefinitionError } from "../../../../api/current-factory-definition";
 import {
@@ -8,235 +6,17 @@ import {
   mockPendingFactoryDocumentSave,
 } from "../../../../testing/factory-document-save-mocks";
 import { staleFactoryVersionTarget } from "../../../../testing/factory-validation-target-fixtures";
+import {
+  createScopedFactoryDocumentSaveQueryClientWrapper,
+  defaultScopedFactoryDocumentSaveRequest,
+  seedScopedFactoryDocumentSaveTestSession,
+} from "../../../../testing/scoped-factory-document-save-test-helpers";
 import * as factoryDocumentSaveHooks from "../../../current-factory-definition/public";
-import { DashboardSessionProvider } from "../../../dashboard/session/dashboard-session-provider";
-import { useDashboardSessionStore } from "../../../dashboard/state/dashboardSessionStore";
-import type { ScopedFactoryDocumentSaveRequest } from "./useScopedFactoryDocumentSave";
 import { useScopedFactoryDocumentSave } from "./useScopedFactoryDocumentSave";
 
-const defaultSaveRequest: ScopedFactoryDocumentSaveRequest = {
-  baseVersion: {
-    logical: "7",
-    physical: "2026-05-23T15:52:00Z",
-  },
-  factory: {
-    name: "Current Factory",
-    workers: [],
-    workstations: [],
-  },
-  scopeKey: "review:transition:Review",
-};
-
 beforeEach(() => {
-  useDashboardSessionStore.setState({ selectedSessionID: "~default" });
+  seedScopedFactoryDocumentSaveTestSession();
   vi.restoreAllMocks();
-});
-
-describe("useScopedFactoryDocumentSave scope isolation", () => {
-  it("clears confirmation, success, warning, and error state when scopeKey changes", async () => {
-    const saveMutation = mockFactoryDocumentSave({
-      mode: "error",
-      rejectedError: new CurrentFactoryDefinitionError("Save failed.", {
-        code: "BAD_REQUEST",
-      }),
-    });
-    vi.spyOn(
-      factoryDocumentSaveHooks,
-      "useFactoryDocumentSave",
-    ).mockReturnValue(saveMutation as never);
-
-    const { rerender, result } = renderHook(
-      ({ scopeKey }) =>
-        useScopedFactoryDocumentSave({
-          fallbackErrorMessage: "Unable to save the active factory.",
-          scopeKey,
-        }),
-      {
-        initialProps: { scopeKey: "review:transition:Review" },
-        wrapper: createQueryClientWrapper(),
-      },
-    );
-
-    act(() => {
-      result.current.beginConfirmation();
-    });
-    expect(result.current.saveState).toEqual({ status: "confirming" });
-
-    await act(async () => {
-      await result.current.confirmSave(defaultSaveRequest);
-    });
-
-    await waitFor(() => {
-      expect(result.current.saveState).toEqual({
-        errorMessage: "Save failed.",
-        status: "error",
-      });
-    });
-
-    rerender({ scopeKey: "worker:reviewer" });
-
-    expect(result.current.saveState).toEqual({ status: "idle" });
-  });
-
-  it("surfaces success on a new scope after rerendering from a previous scope", async () => {
-    const saveMutation = mockFactoryDocumentSave({ mode: "success" });
-    vi.spyOn(
-      factoryDocumentSaveHooks,
-      "useFactoryDocumentSave",
-    ).mockReturnValue(saveMutation as never);
-
-    const scopeBSaveRequest: ScopedFactoryDocumentSaveRequest = {
-      ...defaultSaveRequest,
-      scopeKey: "worker:reviewer",
-    };
-
-    const { rerender, result } = renderHook(
-      ({ scopeKey }) =>
-        useScopedFactoryDocumentSave({
-          fallbackErrorMessage: "Unable to save the active factory.",
-          scopeKey,
-        }),
-      {
-        initialProps: { scopeKey: "review:transition:Review" },
-        wrapper: createQueryClientWrapper(),
-      },
-    );
-
-    rerender({ scopeKey: "worker:reviewer" });
-
-    await act(async () => {
-      await result.current.saveNow(scopeBSaveRequest);
-    });
-
-    await waitFor(() => {
-      expect(result.current.saveState).toEqual({ status: "success" });
-    });
-  });
-
-  it("clears success when the draft becomes dirty again for the current scope", async () => {
-    const saveMutation = mockFactoryDocumentSave({ mode: "success" });
-    vi.spyOn(
-      factoryDocumentSaveHooks,
-      "useFactoryDocumentSave",
-    ).mockReturnValue(saveMutation as never);
-
-    const { rerender, result } = renderHook(
-      ({ isDirty }) =>
-        useScopedFactoryDocumentSave({
-          fallbackErrorMessage: "Unable to save the active factory.",
-          isDirty,
-          scopeKey: "review:transition:Review",
-        }),
-      {
-        initialProps: { isDirty: false },
-        wrapper: createQueryClientWrapper(),
-      },
-    );
-
-    await act(async () => {
-      await result.current.saveNow(defaultSaveRequest);
-    });
-
-    await waitFor(() => {
-      expect(result.current.saveState).toEqual({ status: "success" });
-    });
-
-    rerender({ isDirty: true });
-
-    expect(result.current.saveState).toEqual({ status: "idle" });
-  });
-
-  it("does not apply success state or onSaved when scopeKey changes during an in-flight save", async () => {
-    const pendingSave = mockPendingFactoryDocumentSave();
-    vi.spyOn(
-      factoryDocumentSaveHooks,
-      "useFactoryDocumentSave",
-    ).mockReturnValue(pendingSave.saveMutation as never);
-    const onSaved = vi.fn();
-
-    const { rerender, result } = renderHook(
-      ({ scopeKey }) =>
-        useScopedFactoryDocumentSave({
-          fallbackErrorMessage: "Unable to save the active factory.",
-          scopeKey,
-        }),
-      {
-        initialProps: { scopeKey: "review:transition:Review" },
-        wrapper: createQueryClientWrapper(),
-      },
-    );
-
-    let savePromise: Promise<void> | undefined;
-    await act(async () => {
-      savePromise = result.current.saveNow({
-        ...defaultSaveRequest,
-        onSaved,
-      });
-      await Promise.resolve();
-    });
-
-    expect(result.current.saveState).toEqual({ status: "submitting" });
-
-    rerender({ scopeKey: "worker:reviewer" });
-    expect(result.current.saveState).toEqual({ status: "idle" });
-
-    pendingSave.deferred.resolve({
-      name: "Current Factory",
-      version: {
-        logical: "8",
-        physical: "2026-05-23T15:52:00.001Z",
-      },
-      workers: [],
-      workstations: [],
-    });
-
-    await act(async () => {
-      await savePromise;
-    });
-
-    expect(onSaved).not.toHaveBeenCalled();
-    expect(result.current.saveState).toEqual({ status: "idle" });
-  });
-
-  it("does not apply error state when scopeKey changes during an in-flight save", async () => {
-    const pendingSave = mockPendingFactoryDocumentSave();
-    vi.spyOn(
-      factoryDocumentSaveHooks,
-      "useFactoryDocumentSave",
-    ).mockReturnValue(pendingSave.saveMutation as never);
-
-    const { rerender, result } = renderHook(
-      ({ scopeKey }) =>
-        useScopedFactoryDocumentSave({
-          fallbackErrorMessage: "Unable to save the active factory.",
-          scopeKey,
-        }),
-      {
-        initialProps: { scopeKey: "review:transition:Review" },
-        wrapper: createQueryClientWrapper(),
-      },
-    );
-
-    let savePromise: Promise<void> | undefined;
-    await act(async () => {
-      savePromise = result.current.saveNow(defaultSaveRequest);
-      await Promise.resolve();
-    });
-
-    rerender({ scopeKey: "worker:reviewer" });
-
-    pendingSave.deferred.reject(
-      new CurrentFactoryDefinitionError("Save failed.", {
-        code: "BAD_REQUEST",
-      }),
-    );
-
-    await act(async () => {
-      await savePromise;
-    });
-
-    expect(result.current.saveState).toEqual({ status: "idle" });
-  });
 });
 
 describe("useScopedFactoryDocumentSave in-flight deduplication", () => {
@@ -254,7 +34,7 @@ describe("useScopedFactoryDocumentSave in-flight deduplication", () => {
           fallbackErrorMessage: "Unable to save the active factory.",
           scopeKey: "review:transition:Review",
         }),
-      { wrapper: createQueryClientWrapper() },
+      { wrapper: createScopedFactoryDocumentSaveQueryClientWrapper() },
     );
 
     act(() => {
@@ -263,10 +43,12 @@ describe("useScopedFactoryDocumentSave in-flight deduplication", () => {
 
     let firstSave: Promise<void> | undefined;
     await act(async () => {
-      firstSave = result.current.confirmSave(defaultSaveRequest);
+      firstSave = result.current.confirmSave(
+        defaultScopedFactoryDocumentSaveRequest,
+      );
       await Promise.resolve();
-      await result.current.confirmSave(defaultSaveRequest);
-      await result.current.saveNow(defaultSaveRequest);
+      await result.current.confirmSave(defaultScopedFactoryDocumentSaveRequest);
+      await result.current.saveNow(defaultScopedFactoryDocumentSaveRequest);
     });
 
     expect(saveAsync).toHaveBeenCalledTimes(1);
@@ -309,11 +91,11 @@ describe("useScopedFactoryDocumentSave error normalization", () => {
           fallbackErrorMessage: "Unable to save the active factory.",
           scopeKey: "review:transition:Review",
         }),
-      { wrapper: createQueryClientWrapper() },
+      { wrapper: createScopedFactoryDocumentSaveQueryClientWrapper() },
     );
 
     await act(async () => {
-      await result.current.saveNow(defaultSaveRequest);
+      await result.current.saveNow(defaultScopedFactoryDocumentSaveRequest);
     });
 
     await waitFor(() => {
@@ -367,15 +149,17 @@ describe("useScopedFactoryDocumentSave error normalization", () => {
                 fieldErrors.workerName = error.message;
               }
             }
-            return Object.keys(fieldErrors).length > 0 ? fieldErrors : undefined;
+            return Object.keys(fieldErrors).length > 0
+              ? fieldErrors
+              : undefined;
           },
           scopeKey: "review:transition:Review",
         }),
-      { wrapper: createQueryClientWrapper() },
+      { wrapper: createScopedFactoryDocumentSaveQueryClientWrapper() },
     );
 
     await act(async () => {
-      await result.current.saveNow(defaultSaveRequest);
+      await result.current.saveNow(defaultScopedFactoryDocumentSaveRequest);
     });
 
     await waitFor(() => {
@@ -405,11 +189,11 @@ describe("useScopedFactoryDocumentSave error normalization", () => {
           fallbackErrorMessage: "Unable to save the active factory.",
           scopeKey: "review:transition:Review",
         }),
-      { wrapper: createQueryClientWrapper() },
+      { wrapper: createScopedFactoryDocumentSaveQueryClientWrapper() },
     );
 
     await act(async () => {
-      await result.current.saveNow(defaultSaveRequest);
+      await result.current.saveNow(defaultScopedFactoryDocumentSaveRequest);
     });
 
     await waitFor(() => {
@@ -435,7 +219,7 @@ describe("useScopedFactoryDocumentSave confirm mode", () => {
           fallbackErrorMessage: "Unable to save the active factory.",
           scopeKey: "review:transition:Review",
         }),
-      { wrapper: createQueryClientWrapper() },
+      { wrapper: createScopedFactoryDocumentSaveQueryClientWrapper() },
     );
 
     act(() => {
@@ -464,7 +248,7 @@ describe("useScopedFactoryDocumentSave confirm mode", () => {
           fallbackErrorMessage: "Unable to save the active factory.",
           scopeKey: "review:transition:Review",
         }),
-      { wrapper: createQueryClientWrapper() },
+      { wrapper: createScopedFactoryDocumentSaveQueryClientWrapper() },
     );
 
     act(() => {
@@ -473,7 +257,7 @@ describe("useScopedFactoryDocumentSave confirm mode", () => {
 
     await act(async () => {
       await result.current.confirmSave({
-        ...defaultSaveRequest,
+        ...defaultScopedFactoryDocumentSaveRequest,
         onSaved,
       });
     });
@@ -483,8 +267,8 @@ describe("useScopedFactoryDocumentSave confirm mode", () => {
     });
     expect(onSaved).toHaveBeenCalledTimes(1);
     expect(saveAsync).toHaveBeenCalledWith({
-      baseVersion: defaultSaveRequest.baseVersion,
-      factory: defaultSaveRequest.factory,
+      baseVersion: defaultScopedFactoryDocumentSaveRequest.baseVersion,
+      factory: defaultScopedFactoryDocumentSaveRequest.factory,
     });
   });
 
@@ -511,7 +295,7 @@ describe("useScopedFactoryDocumentSave confirm mode", () => {
           fallbackErrorMessage: "Unable to save the active factory.",
           scopeKey: "review:transition:Review",
         }),
-      { wrapper: createQueryClientWrapper() },
+      { wrapper: createScopedFactoryDocumentSaveQueryClientWrapper() },
     );
 
     act(() => {
@@ -519,7 +303,7 @@ describe("useScopedFactoryDocumentSave confirm mode", () => {
     });
 
     await act(async () => {
-      await result.current.confirmSave(defaultSaveRequest);
+      await result.current.confirmSave(defaultScopedFactoryDocumentSaveRequest);
     });
 
     await waitFor(() => {
@@ -528,17 +312,3 @@ describe("useScopedFactoryDocumentSave confirm mode", () => {
     expect(result.current.isPending).toBe(false);
   });
 });
-
-function createQueryClientWrapper() {
-  const queryClient = new QueryClient({
-    defaultOptions: { mutations: { retry: false }, queries: { retry: false } },
-  });
-
-  return function QueryClientWrapper({ children }: { children: ReactNode }) {
-    return (
-      <QueryClientProvider client={queryClient}>
-        <DashboardSessionProvider>{children}</DashboardSessionProvider>
-      </QueryClientProvider>
-    );
-  };
-}
