+++ ui/src/features/workflow-activity/hooks/react-flow-current-activity-card-editor.tsx
@@ -1,12 +1,11 @@
-import { useMemo, useRef, useState } from "react";
+import { useState } from "react";
 
 import type { DashboardSnapshot } from "../../../api/dashboard/types";
 import type { FactoryGraphEditorTool } from "../../factory-graph-editor/components/factory-graph-editor-controls";
-import type { EditableFactoryGraphViewModel } from "../../factory-graph-editor/hooks/use-editable-factory-graph-types";
-import { buildDraftAppliedFactoryDefinition } from "../../factory-graph-editor/lib/factory-graph-draft-apply";
-import { useFactoryValidation } from "../../factory-graph-editor/hooks/use-factory-validation";
-import { buildCurrentActivityGraphEditorValue } from "./react-flow-current-activity-card-editor-value";
+import { useDraftAppliedFactoryValidation } from "./react-flow-current-activity-card-editor-draft-validation";
 import { useCurrentActivityEditableGraph } from "./react-flow-current-activity-card-editor-editable-graph";
+import { useGraphEditorLeaveEditorBridge } from "./react-flow-current-activity-card-editor-save-flow";
+import { buildCurrentActivityGraphEditorValue } from "./react-flow-current-activity-card-editor-value";
 import { useGraphEditorControllers } from "./use-graph-editor-controllers";
 import { useGraphEditorSaveFlow } from "./use-graph-editor-save-flow";
 import { useGraphEditorSession } from "./use-graph-editor-session";
@@ -18,10 +17,7 @@ export function useCurrentActivityGraphEditor(
 ) {
   const [editorMode, setEditorMode] = useState(false);
   const [activeTool, setActiveTool] = useState<FactoryGraphEditorTool>(null);
-  const leaveEditorRef = useRef<() => void>(() => {});
-  const setIsConfirmingLeaveEditorRef = useRef<(open: boolean) => void>(
-    () => {},
-  );
+  const leaveEditorBridge = useGraphEditorLeaveEditorBridge();
   const { currentFactoryQuery, editableGraph, saveEditableDefinition } =
     useCurrentActivityEditableGraph({
       editorMode,
@@ -40,10 +36,7 @@ export function useCurrentActivityGraphEditor(
     editableDefinitionQuery: currentFactoryQuery,
     editorMode,
     locale,
-    onAttemptLeaveEditor: () => setIsConfirmingLeaveEditorRef.current(true),
-    onLeaveEditor: () => {
-      leaveEditorRef.current();
-    },
+    ...leaveEditorBridge.sessionCallbacks,
     projectedFactory: snapshot.factory,
     saveEditableDefinition,
     setActiveTool,
@@ -79,8 +72,7 @@ export function useCurrentActivityGraphEditor(
       setPendingRemovalNodeId: controllers.setPendingRemovalNodeId,
     },
   });
-  setIsConfirmingLeaveEditorRef.current = saveFlow.setIsConfirmingLeaveEditor;
-  leaveEditorRef.current = saveFlow.leaveEditor;
+  leaveEditorBridge.bindSaveFlow(saveFlow);
 
   return buildCurrentActivityGraphEditorValue({
     activeTool,
@@ -127,21 +119,4 @@ export function useCurrentActivityGraphEditor(
     setPendingRemovalNodeId: controllers.setPendingRemovalNodeId,
     structuralValidation,
   });
-}
-
-function useDraftAppliedFactoryValidation(
-  draftState: EditableFactoryGraphViewModel["draftState"],
-  editorMode: boolean,
-) {
-  const draftAppliedFactoryDefinition = useMemo(() => {
-    const baseDocument =
-      draftState.latestDocument ?? draftState.baseDocument ?? null;
-    if (!baseDocument) {
-      return null;
-    }
-
-    return buildDraftAppliedFactoryDefinition(baseDocument, draftState.draft);
-  }, [draftState.baseDocument, draftState.draft, draftState.latestDocument]);
-
-  return useFactoryValidation(draftAppliedFactoryDefinition, editorMode);
 }
